Задание к блоку 5
